Added traits
