+++ src/S3/Result/CreateBucketOutputTrait.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace AsyncAws\S3\Result;
+
+use Symfony\Contracts\HttpClient\ResponseInterface;
+
+trait CreateBucketOutputTrait
+{
+    protected function populateFromResponse(ResponseInterface $response): void
+    {
+        $data = new \SimpleXMLElement($response->getContent(false));
+
+        // TODO Verify correctness
+        $this->Location = $data->Location;
+    }
+}
