+++ src/S3/Result/PutObjectOutput.php
@@ -3,10 +3,11 @@
 namespace AsyncAws\S3\Result;
 
 use AsyncAws\Aws\Result;
-use Symfony\Contracts\HttpClient\ResponseInterface;
 
 class PutObjectOutput extends Result
 {
+    use PutObjectOutputTrait;
+
     private $Expiration;
 
     private $ETag;
@@ -24,22 +25,6 @@ class PutObjectOutput extends Result
     private $SSEKMSEncryptionContext;
 
     private $RequestCharged;
-
-    protected function populateFromResponse(ResponseInterface $response): void
-    {
-        $data = new \SimpleXMLElement($response->getContent(false));
-
-        // TODO Verify correctness
-        $this->Expiration = $data->Expiration;
-        $this->ETag = $data->ETag;
-        $this->ServerSideEncryption = $data->ServerSideEncryption;
-        $this->VersionId = $data->VersionId;
-        $this->SSECustomerAlgorithm = $data->SSECustomerAlgorithm;
-        $this->SSECustomerKeyMD5 = $data->SSECustomerKeyMD5;
-        $this->SSEKMSKeyId = $data->SSEKMSKeyId;
-        $this->SSEKMSEncryptionContext = $data->SSEKMSEncryptionContext;
-        $this->RequestCharged = $data->RequestCharged;
-    }
 
     public function getExpiration(): string
     {
